refactor: Make repository's getGenres function asynchronous.
The getGenres() function in the repository layer was previously synchronous. This coupled the ViewModel to a specific implementation and would break the ViewModel if the source ever change to a network or database call.
By adding the 'suspend' modifier to the repository and use case, we decouple the ViewModel from the data origin